backend/requirements.txt
# TimeValue 个人资产管理系统 - Python依赖
# Powered by 孚普科技（北京）有限公司
# https://github.com/fupukeji/timevalue

# ===========================
# Web框架核心
# ===========================
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1

# ===========================
# 数据库
# ===========================
SQLAlchemy==2.0.23
PyMySQL==1.1.0
cryptography==41.0.7

# ===========================
# 密码加密
# ===========================
Werkzeug==3.0.1

# ===========================
# 环境变量管理
# ===========================
python-dotenv==1.0.0

# ===========================
# HTTP请求库
# ===========================
requests==2.31.0

# ===========================
# LangGraph工作流引擎
# ===========================
langgraph==0.2.0
langchain-core>=0.2.27,<0.3.0

# ===========================
# 日期时间处理
# ===========================
python-dateutil==2.8.2

# ===========================
# 工具库
# ===========================
# 已包含在Python标准库中:
# - json
# - base64
# - os
# - secrets
# - string
# - pathlib
# - datetime
# - decimal
# - logging
# - asyncio
# - typing

# ===========================
# 生产环境建议
# ===========================
# 部署时建议添加以下依赖:
# gunicorn==21.2.0  # WSGI服务器
# python-json-logger==2.0.7  # 日志格式化